+++ src/main/java/com/raoulvdberge/refinedstorageaddons/RSAddonsItems.java
@@ -1,7 +1,9 @@
 package com.raoulvdberge.refinedstorageaddons;
 
+import com.raoulvdberge.refinedstorageaddons.item.ItemNetworkPicker;
 import com.raoulvdberge.refinedstorageaddons.item.ItemWirelessCraftingGrid;
 
 public final class RSAddonsItems {
     public static final ItemWirelessCraftingGrid WIRELESS_CRAFTING_GRID = new ItemWirelessCraftingGrid();
+    public static final ItemNetworkPicker NETWORK_PICKER = new ItemNetworkPicker();
 }