+++ src/main/java/com/raoulvdberge/refinedstorageaddons/proxy/ProxyCommon.java
@@ -4,10 +4,11 @@ import com.raoulvdberge.refinedstorage.apiimpl.network.node.NetworkNode;
 import com.raoulvdberge.refinedstorageaddons.RSAddons;
 import com.raoulvdberge.refinedstorageaddons.RSAddonsBlocks;
 import com.raoulvdberge.refinedstorageaddons.RSAddonsItems;
+import com.raoulvdberge.refinedstorageaddons.apiimpl.network.grid.wireless.WirelessGridFactoryWirelessCraftingGrid;
 import com.raoulvdberge.refinedstorageaddons.apiimpl.network.node.NetworkNodeInfiniteWirelessTransmitter;
 import com.raoulvdberge.refinedstorageaddons.gui.GuiHandler;
 import com.raoulvdberge.refinedstorageaddons.item.WirelessCraftingGrid;
-import com.raoulvdberge.refinedstorageaddons.item.WirelessGridFactoryWirelessCraftingGrid;
+import com.raoulvdberge.refinedstorageaddons.network.MessagePickBlock;
 import com.raoulvdberge.refinedstorageaddons.tile.TileInfiniteWirelessTransmitter;
 import net.minecraft.block.Block;
 import net.minecraft.item.Item;
@@ -19,6 +20,7 @@ import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 import net.minecraftforge.fml.common.network.NetworkRegistry;
 import net.minecraftforge.fml.common.registry.GameRegistry;
+import net.minecraftforge.fml.relauncher.Side;
 
 public class ProxyCommon {
     public void preInit(FMLPreInitializationEvent e) {
@@ -38,6 +40,8 @@ public class ProxyCommon {
 
         GameRegistry.registerTileEntity(TileInfiniteWirelessTransmitter.class, RSAddons.ID + ":infinite_wireless_transmitter");
 
+        RSAddons.INSTANCE.network.registerMessage(MessagePickBlock.class, MessagePickBlock.class, 0, Side.SERVER);
+
         NetworkRegistry.INSTANCE.registerGuiHandler(RSAddons.INSTANCE, new GuiHandler());
     }
 
@@ -53,6 +57,7 @@ public class ProxyCommon {
     @SubscribeEvent
     public void registerItems(RegistryEvent.Register<Item> e) {
         e.getRegistry().register(RSAddonsItems.WIRELESS_CRAFTING_GRID);
+        e.getRegistry().register(RSAddonsItems.NETWORK_PICKER);
         e.getRegistry().register(RSAddonsBlocks.INFINITE_WIRELESS_TRANSMITTER.createItem());
     }
 }