+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 0.2
 - Updated Forge to 2493 (MC 1.12.2) (raoulvdberge)
 - Added Infinite Wireless Transmitter (can be configured) (raoulvdberge)
+- Added Network Picker (raoulvdberge)
 
 ### 0.1.1
 - Refined Storage 1.5.13 compatibility (raoulvdberge)